refactor(web): update Dropdown component main filename
