+++ apps/web/src/components/Dropdown/components/Option.tsx
@@ -1,6 +1,6 @@
 import { ListboxOption } from "@headlessui/react";
 
-import type { Option as OptionProps } from "../Dropdown";
+import type { Option as OptionProps } from "..";
 
 export const Option: React.FC<OptionProps> = function (props) {
   return (
